refactored and reorganized
