+++ ConsoleClient/BaseOptions.cs
