+++ ConsoleClient/CloudSamples/sample of 100 rects with minimum size=2 and start size=150(1).jpg
