+++ ConsoleClient/Validators/ColorValidator.cs
