+++ ConsoleClient/Validators/StringValidatorAttribute.cs
