+++ ConsoleClient/packages.config
@@ -3,8 +3,6 @@
   <package id="Autofac" version="6.3.0" targetFramework="net48" />
   <package id="CommandLineParser" version="2.9.0-preview1" targetFramework="net48" />
   <package id="Microsoft.Bcl.AsyncInterfaces" version="6.0.0" targetFramework="net48" />
-  <package id="Microsoft.Extensions.DependencyInjection" version="6.0.0" targetFramework="net48" />
-  <package id="Microsoft.Extensions.DependencyInjection.Abstractions" version="6.0.0" targetFramework="net48" />
   <package id="Microsoft.NETFramework.ReferenceAssemblies.net48" version="1.0.0" targetFramework="net48" developmentDependency="true" />
   <package id="System.Buffers" version="4.5.1" targetFramework="net48" />
   <package id="System.Diagnostics.DiagnosticSource" version="4.7.1" targetFramework="net48" />
