+++ GuiClient/GuiClient.csproj
@@ -7,7 +7,7 @@
         <ProjectGuid>{AF87C9D3-28D3-4141-A921-F0CFC0302D37}</ProjectGuid>
         <OutputType>WinExe</OutputType>
         <RootNamespace>DesktopClient</RootNamespace>
-        <AssemblyName>DesktopClient</AssemblyName>
+        <AssemblyName>GuiClient</AssemblyName>
         <TargetFrameworkVersion>v4.8</TargetFrameworkVersion>
         <FileAlignment>512</FileAlignment>
         <AutoGenerateBindingRedirects>true</AutoGenerateBindingRedirects>
