+++ GuiClient/Properties/Resources.resx
