+++ TagsCloudVisualization/Printing/IRectanglesReCalculator.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
 using System.Drawing;
 
-namespace TagsCloudVisualization
+namespace TagsCloudVisualization.Printing
 {
     public interface IRectanglesReCalculator
     {
