+++ TagsCloudVisualization/Printing/RandomColorScheme.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Drawing;
 
-namespace TagsCloudVisualization
+namespace TagsCloudVisualization.Printing
 {
     public class RandomColorScheme : IColorScheme
     {
