+++ TagsCloudVisualization/Statistics/IWordStatisticsToSizeConverter.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 
-namespace TagsCloudVisualization
+namespace TagsCloudVisualization.Statistics
 {
     public interface IWordStatisticsToSizeConverter
     {
