+++ TagsCloudVisualization/Statistics/LiteraryWordsStatistics.cs
@@ -1,0 +1,11 @@
+﻿using TagsCloudVisualization.WordProcessors;
+
+namespace TagsCloudVisualization.Statistics
+{
+    public class LiteraryWordsStatistics : BaseWordsStatistics
+    {
+        public LiteraryWordsStatistics(LiteraryTextProcessor textProcessor) : base(textProcessor)
+        {
+        }
+    }
+}
