+++ TagsCloudVisualization/Statistics/OneWordByLineStatistics.cs
@@ -1,0 +1,11 @@
+﻿using TagsCloudVisualization.WordProcessors;
+
+namespace TagsCloudVisualization.Statistics
+{
+    public class OneWordByLineStatistics : BaseWordsStatistics
+    {
+        public OneWordByLineStatistics(OneWordByLineProcessor textProcessor) : base(textProcessor)
+        {
+        }
+    }
+}
