--- TagsCloudVisualization/ignore.txt
@@ -1,12 +1,0 @@
-в
-на
-что
-ты
-я
-вы
-он
-на
-из
-от
-за
-для
