+++ TagsCloudVisualization/packages.config
@@ -5,8 +5,6 @@
   <package id="DocumentFormat.OpenXml" version="2.14.0" targetFramework="net48" />
   <package id="FluentAssertions" version="6.2.0" targetFramework="net48" />
   <package id="Microsoft.Bcl.AsyncInterfaces" version="6.0.0" targetFramework="net48" />
-  <package id="Microsoft.Extensions.DependencyInjection" version="6.0.0" targetFramework="net48" />
-  <package id="Microsoft.Extensions.DependencyInjection.Abstractions" version="6.0.0" targetFramework="net48" />
   <package id="Microsoft.NETFramework.ReferenceAssemblies.net48" version="1.0.0" targetFramework="net48" developmentDependency="true" />
   <package id="Microsoft.Office.Interop.Word" version="15.0.4797.1003" targetFramework="net48" />
   <package id="NUnit" version="3.13.2" targetFramework="net48" />
