+++ TagsCloudVisualizationTest/packages.config
@@ -5,8 +5,6 @@
   <package id="FakeItEasy" version="7.2.0" targetFramework="net48" />
   <package id="FluentAssertions" version="6.2.0" targetFramework="net48" />
   <package id="Microsoft.Bcl.AsyncInterfaces" version="6.0.0" targetFramework="net48" />
-  <package id="Microsoft.Extensions.DependencyInjection" version="6.0.0" targetFramework="net48" />
-  <package id="Microsoft.Extensions.DependencyInjection.Abstractions" version="6.0.0" targetFramework="net48" />
   <package id="NUnit" version="3.5.0" targetFramework="net45" />
   <package id="System.Buffers" version="4.5.1" targetFramework="net48" />
   <package id="System.Diagnostics.DiagnosticSource" version="4.7.1" targetFramework="net48" />
